+++ tests/node_functions.py
@@ -1,12 +1,13 @@
 from primitiv import Device
-from primitiv import tensor_functions as tF
+from primitiv import Graph
+from primitiv import functions as F
 from primitiv.devices import Naive
 
 import numpy as np
 import unittest
 
 
-class TensorOperatorsTest(unittest.TestCase):
+class NodeFunctionsTest(unittest.TestCase):
 
     @classmethod
     def setUpClass(cls):
@@ -18,96 +19,75 @@ class TensorOperatorsTest(unittest.TestCase):
 
     def setUp(self):
         self.device = Naive()
+        self.graph = Graph()
         Device.set_default(self.device)
+        Graph.set_default(self.graph)
         self.a = np.array([[1, 2], [3, 4]], np.float32)
         self.b = np.array([[1, 1], [4, 8]], np.float32)
 
     def tearDown(self):
         pass
 
-    def test_tensor_pos(self):
-        x = tF.input(self.a)
-        y = tF.input(self.b)
+    def test_node_pos(self):
+        x = F.input(self.a)
+        y = F.input(self.b)
         self.assertTrue(((+x).to_ndarrays()[0] == self.a).all())
 
-    def test_tensor_neg(self):
-        x = tF.input(self.a)
-        y = tF.input(self.b)
+    def test_node_neg(self):
+        x = F.input(self.a)
+        y = F.input(self.b)
         self.assertTrue(((-x).to_ndarrays()[0] == -self.a).all())
 
-    def test_tensor_add(self):
-        x = tF.input(self.a)
-        y = tF.input(self.b)
+    def test_node_add(self):
+        x = F.input(self.a)
+        y = F.input(self.b)
         self.assertTrue(((x + y).to_ndarrays()[0] == np.array([[2, 3], [7, 12]])).all())
         self.assertTrue(((x + 2).to_ndarrays()[0] == np.array([[3, 4], [5, 6]])).all())
         self.assertTrue(((2 + x).to_ndarrays()[0] == np.array([[3, 4], [5, 6]])).all())
 
-    def test_tensor_sub(self):
-        x = tF.input(self.a)
-        y = tF.input(self.b)
+    def test_node_sub(self):
+        x = F.input(self.a)
+        y = F.input(self.b)
         self.assertTrue(((x - y).to_ndarrays()[0] == np.array([[0, 1], [-1, -4]])).all())
         self.assertTrue(((x - 2).to_ndarrays()[0] == np.array([[-1, 0], [1, 2]])).all())
         self.assertTrue(((2 - x).to_ndarrays()[0] == np.array([[1, 0], [-1, -2]])).all())
 
-    def test_tensor_mul(self):
-        x = tF.input(self.a)
-        y = tF.input(self.b)
+    def test_node_mul(self):
+        x = F.input(self.a)
+        y = F.input(self.b)
         self.assertTrue(((x * y).to_ndarrays()[0] == np.array([[1, 2], [12, 32]])).all())
         self.assertTrue(((x * 2).to_ndarrays()[0] == np.array([[2, 4], [6, 8]])).all())
         self.assertTrue(((2 * x).to_ndarrays()[0] == np.array([[2, 4], [6, 8]])).all())
 
-    def test_tensor_matmul(self):
-        x = tF.input(self.a)
-        y = tF.input(self.b)
+    def test_node_matmul(self):
+        x = F.input(self.a)
+        y = F.input(self.b)
         self.assertTrue(((x @ y).to_ndarrays()[0] == np.array([[9, 17], [19, 35]])).all())
         self.assertRaises(TypeError, lambda: x @ 2)
         self.assertRaises(TypeError, lambda: 2 @ x)
 
-    def test_tensor_truediv(self):
-        x = tF.input(self.a)
-        y = tF.input(self.b)
+    def test_node_truediv(self):
+        x = F.input(self.a)
+        y = F.input(self.b)
         self.assertTrue(((x / y).to_ndarrays()[0] == np.array([[1, 2], [0.75, 0.5]])).all())
         self.assertTrue(((x / 2).to_ndarrays()[0] == np.array([[0.5, 1], [1.5, 2]])).all())
         self.assertTrue(((2 / y).to_ndarrays()[0] == np.array([[2, 2], [0.5, 0.25]])).all())
 
-    def test_tensor_pow(self):
-        x = tF.input(self.a)
-        y = tF.input(self.b)
+    def test_node_pow(self):
+        x = F.input(self.a)
+        y = F.input(self.b)
         self.assertTrue(np.isclose((x ** y).to_ndarrays()[0], np.array([[1, 2], [81, 65536]])).all())
         self.assertTrue(np.isclose((x ** 2).to_ndarrays()[0], np.array([[1, 4], [9, 16]])).all())
         self.assertTrue(np.isclose((2 ** x).to_ndarrays()[0], np.array([[2, 4], [8, 16]])).all())
         self.assertTrue(np.isclose((x ** -2).to_ndarrays()[0], np.array([[1, 1/4], [1/9, 1/16]])).all())
         input_arr = np.array([1, -1, 3, -3, 5, -5])
-        x = tF.input(input_arr)
+        x = F.input(input_arr)
         self.assertTrue(((x ** 6).to_ndarrays()[0] == np.array([1, 1, 729, 729, 15625, 15625])).all())
         self.assertTrue(((x ** 9).to_ndarrays()[0] == np.array([1, -1, 19683, -19683, 1953125, -1953125])).all())
         input_arr = np.array([1, -1])
-        x = tF.input(input_arr)
+        x = F.input(input_arr)
         self.assertTrue(((x ** 0x7fffffff).to_ndarrays()[0] == np.array([1, -1])).all())
         self.assertTrue(((x ** -0x80000000).to_ndarrays()[0] == np.array([1, 1])).all())
         self.assertTrue(np.isnan((x ** 0x80000000).to_ndarrays()[0]).any())
         self.assertTrue(np.isnan((x ** -0x80000001).to_ndarrays()[0]).any())
         self.assertRaises(TypeError, lambda: pow(x, y, 2))
-
-    def test_tensor_iadd(self):
-        x = tF.input(self.a)
-        y = tF.input(self.b)
-        x_tmp = x
-        x += y
-        self.assertIs(x, x_tmp)
-        self.assertTrue((x.to_ndarrays()[0] == np.array([[2, 3], [7, 12]])).all())
-
-    def test_tensor_isub(self):
-        x = tF.input(self.a)
-        y = tF.input(self.b)
-        x_tmp = x
-        x -= y
-        self.assertIs(x, x_tmp)
-        self.assertTrue((x.to_ndarrays()[0] == np.array([[0, 1], [-1, -4]])).all())
-
-    def test_tensor_imul(self):
-        x = tF.input(self.a)
-        x_tmp = x
-        x *= 2
-        self.assertIs(x, x_tmp)
-        self.assertTrue((x.to_ndarrays()[0] == np.array([[2, 4], [6, 8]])).all())
